Deduplicate cockpit shared helpers
